Moved isAccount function into account utilities
